requirements.txt
scrapy
selenium
beautifulsoup4
lxml
requests
pytest
flask
python-docx
fpdf
pandas
openpyxl